refactor: move the TaskSubtask sub-components
