+++ wondrous-app/components/Common/TaskSubtask/TaskSubtaskHeader/index.tsx
@@ -1,0 +1,1 @@
+export { TaskSubtaskHeader as default } from './TaskSubtaskHeader';
