+++ wondrous-app/components/Common/TaskSubtask/TaskSubtasks.tsx
@@ -1,0 +1,14 @@
+import { Subtask } from './styles';
+import TaskSubtaskHeader from './TaskSubtaskHeader';
+import TaskSubtaskList from './TaskSubtaskList';
+
+export const TaskSubtasks = ({ taskId, permissions, parentTask }) => {
+  return (
+    <>
+      <Subtask>
+        {taskId && <TaskSubtaskHeader taskId={taskId} permissions={permissions} parentTask={parentTask} />}
+        {taskId && <TaskSubtaskList taskId={taskId} />}
+      </Subtask>
+    </>
+  );
+};
